apps/engine/pyproject.toml
[project]
name = "engine"
version = "0.2.0"
description = "ClearAudio - Audio processing engine powered by SAM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "vercel-blob>=0.4.2",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "fal-client>=0.5.6",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
